Flavor importer #65

+++ database/migrations/2019_12_17_155004_create_plans_table.php
@@ -18,7 +18,7 @@ class CreatePlansTable extends Migration
             $table->string('remote_id');
             $table->string('name');
             $table->integer('disk');
-            $table->integer('ram');
+            $table->float('ram');
             $table->integer('vcpu');
             $table->float('hourly_price');
             $table->timestamps();
